+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+* Annotator can now swap up/down marking mode in situ (providing no mark-changing rubrics present). Menu item or F2 triggers this.
 
 ### Changed
 * Rubrics have had significant change: they are now shared between users and can be grouped into "panes" within the user interface.